+++ lib/bible_parser_flutter.dart
@@ -2,5 +2,6 @@ export 'src/bible_parser.dart';
 export 'src/book.dart';
 export 'src/chapter.dart';
 export 'src/verse.dart';
+export 'src/text_segment.dart';
 export 'src/errors.dart';
 export 'src/bible_repository.dart';